Switch to 16 bit floats on Mac!

+++ stablediffusion/sd_generator.py
@@ -36,7 +36,8 @@ def detect_platform():
     cpu = {"name": "CPU", "device":"cpu", "size":torch.float32, "attention_slicing":False}
     graphcore = {"name": "Graphcore", "device":"ipu", "size":torch.float16, "attention_slicing":True}
     nvidia = {"name": "Nvidia", "device":"cuda", "size":torch.float16, "attention_slicing":False}
-    metal = {"name": "Apple Metal", "device":"mps", "size":torch.float32, "attention_slicing":False}
+    #metal = {"name": "Apple Metal", "device":"mps", "size":torch.float32, "attention_slicing":False} # use this version on an "older" Mac pre OS 14
+    metal = {"name": "Apple Metal", "device":"mps", "size":torch.float16, "attention_slicing":False}
 
     r = cpu
     try:
